Document last_active_segid and preskip_segid.
Change the type of preskip_segid to uint8_t to suggest it is a boolean.

+++ av1/common/seg_common.h
@@ -44,8 +44,12 @@ struct segmentation {
 
   int16_t feature_data[MAX_SEGMENTS][SEG_LVL_MAX];
   unsigned int feature_mask[MAX_SEGMENTS];
-  int last_active_segid;
-  int preskip_segid;
+  int last_active_segid;  // The highest numbered segment id that has some
+                          // enabled feature.
+  uint8_t preskip_segid;  // Whether the segment id will be read before the
+                          // skip syntax element.
+                          // 1: the segment id will be read first.
+                          // 0: the skip syntax element will be read first.
 };
 
 struct segmentation_probs {
